shreyas_API/requirements.txt
django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3